Refactor download, and database update
Separate administrator panel to be outside of the main flask app,
as it performs database operations that may leave the db in an
undefined state if something goes awry. We dont want to lose the admin
panel in that instance.

The admin panel also has a very light version of the login mechanism,
and it's entirely stateless - it only looks at the current users' google id